# cgosqlite: get column types during Step